requirements.txt
# LLM面试模拟系统 - Enhanced版依赖包
# 基于DeepSeek + BGE + Qdrant技术栈

# 核心框架
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-core>=0.1.0

# LLM和API
openai>=1.0.0
python-dotenv>=1.0.0

# 向量和嵌入
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
faiss-cpu>=1.7.4
huggingface-hub>=0.16.0

# 向量数据库 (可选)
qdrant-client>=1.6.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# 文档处理
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# Web scraping (用于文档加载)
selenium>=4.15.0

# 开发工具
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0
flake8>=6.0.0

# 系统工具
tqdm>=4.66.0
pathlib>=1.0.1

# 可选：如果需要GPU加速
# torch-audio  # 取消注释如果需要音频处理
# accelerate   # 取消注释如果需要模型加速

# 可选：如果需要更多文档格式支持
# pypdf2       # PDF支持
# python-docx  # Word文档支持
# openpyxl     # Excel支持

# 注意事项：
# 1. torch版本请根据你的CUDA版本选择对应版本
# 2. 如果遇到依赖冲突，可以创建新的虚拟环境
# 3. 某些包可能需要特定的Python版本（推荐Python 3.9-3.11）

# 安装命令示例：
# pip install -r requirements.txt
# 
# 或分步安装：
# pip install langchain langchain-community langchain-openai
# pip install openai python-dotenv sentence-transformers
# pip install faiss-cpu qdrant-client numpy pandas
